Fix type conversion warnings, suppress 4099 (debug pdb missing), .gitignore
Player::crntFirePattern should be size_t
to_float when passing physics values (doubles) to DrawNode
Fix color for IllusionDialDagger bullet
Fix use of double literals for Color4F
Ignore incredibuild files

+++ Classes/objects/Player.hpp
@@ -94,7 +94,7 @@ public:
     
 	FirePattern* getFirePattern();
 
-	bool trySetFirePattern(int idx);
+	bool trySetFirePattern(size_t idx);
 	bool trySetFirePatternNext();
 	bool trySetFirePatternPrevious();
 
@@ -123,7 +123,7 @@ protected:
 	boost::rational<int> interactCooldown = 0;
 
 	vector<unique_ptr<FirePattern>> firePatterns;
-	int crntFirePattern = 0;
+	size_t crntFirePattern = 0;
 
 	SpellDesc* equippedSpell = nullptr;
